improve id signature binary format

+++ compiler/ir/serialization.common/src/org/jetbrains/kotlin/backend/common/serialization/proto/ActualOrBuilder.java
@@ -8,20 +8,36 @@ public interface ActualOrBuilder extends
     org.jetbrains.kotlin.protobuf.MessageLiteOrBuilder {
 
   /**
-   * <code>required int32 actual_symbol = 1;</code>
+   * <code>required int64 actual_symbol = 1;</code>
+   *
+   * <pre>
+   *    required int32 actual_symbol = 1;
+   * </pre>
    */
   boolean hasActualSymbol();
   /**
-   * <code>required int32 actual_symbol = 1;</code>
+   * <code>required int64 actual_symbol = 1;</code>
+   *
+   * <pre>
+   *    required int32 actual_symbol = 1;
+   * </pre>
    */
-  int getActualSymbol();
+  long getActualSymbol();
 
   /**
-   * <code>required int32 expect_symbol = 2;</code>
+   * <code>required int64 expect_symbol = 2;</code>
+   *
+   * <pre>
+   *    required int32 expect_symbol = 2;
+   * </pre>
    */
   boolean hasExpectSymbol();
   /**
-   * <code>required int32 expect_symbol = 2;</code>
+   * <code>required int64 expect_symbol = 2;</code>
+   *
+   * <pre>
+   *    required int32 expect_symbol = 2;
+   * </pre>
    */
-  int getExpectSymbol();
+  long getExpectSymbol();
 }
