+++ compiler/ir/serialization.common/src/org/jetbrains/kotlin/backend/common/serialization/proto/IrClassReferenceOrBuilder.java
@@ -8,13 +8,21 @@ public interface IrClassReferenceOrBuilder extends
     org.jetbrains.kotlin.protobuf.MessageLiteOrBuilder {
 
   /**
-   * <code>required int32 class_symbol = 1;</code>
+   * <code>required int64 class_symbol = 1;</code>
+   *
+   * <pre>
+   *  required int32 class_symbol = 1;
+   * </pre>
    */
   boolean hasClassSymbol();
   /**
-   * <code>required int32 class_symbol = 1;</code>
+   * <code>required int64 class_symbol = 1;</code>
+   *
+   * <pre>
+   *  required int32 class_symbol = 1;
+   * </pre>
    */
-  int getClassSymbol();
+  long getClassSymbol();
 
   /**
    * <code>required int32 class_type = 2;</code>
