+++ compiler/ir/serialization.common/src/org/jetbrains/kotlin/backend/common/serialization/proto/IrPropertyReferenceOrBuilder.java
@@ -8,31 +8,55 @@ public interface IrPropertyReferenceOrBuilder extends
     org.jetbrains.kotlin.protobuf.MessageLiteOrBuilder {
 
   /**
-   * <code>optional int32 field = 1;</code>
+   * <code>optional int64 field = 1;</code>
+   *
+   * <pre>
+   *  optional int32 field = 1;
+   * </pre>
    */
   boolean hasField();
   /**
-   * <code>optional int32 field = 1;</code>
+   * <code>optional int64 field = 1;</code>
+   *
+   * <pre>
+   *  optional int32 field = 1;
+   * </pre>
    */
-  int getField();
+  long getField();
 
   /**
-   * <code>optional int32 getter = 2;</code>
+   * <code>optional int64 getter = 2;</code>
+   *
+   * <pre>
+   *  optional int32 getter = 2;
+   * </pre>
    */
   boolean hasGetter();
   /**
-   * <code>optional int32 getter = 2;</code>
+   * <code>optional int64 getter = 2;</code>
+   *
+   * <pre>
+   *  optional int32 getter = 2;
+   * </pre>
    */
-  int getGetter();
+  long getGetter();
 
   /**
-   * <code>optional int32 setter = 3;</code>
+   * <code>optional int64 setter = 3;</code>
+   *
+   * <pre>
+   *  optional int32 setter = 3;
+   * </pre>
    */
   boolean hasSetter();
   /**
-   * <code>optional int32 setter = 3;</code>
+   * <code>optional int64 setter = 3;</code>
+   *
+   * <pre>
+   *  optional int32 setter = 3;
+   * </pre>
    */
-  int getSetter();
+  long getSetter();
 
   /**
    * <code>optional .org.jetbrains.kotlin.backend.common.serialization.proto.IrStatementOrigin origin = 4;</code>
@@ -53,11 +77,19 @@ public interface IrPropertyReferenceOrBuilder extends
   org.jetbrains.kotlin.backend.common.serialization.proto.MemberAccessCommon getMemberAccess();
 
   /**
-   * <code>required int32 symbol = 6;</code>
+   * <code>required int64 symbol = 6;</code>
+   *
+   * <pre>
+   *  required int32 symbol = 6;
+   * </pre>
    */
   boolean hasSymbol();
   /**
-   * <code>required int32 symbol = 6;</code>
+   * <code>required int64 symbol = 6;</code>
+   *
+   * <pre>
+   *  required int32 symbol = 6;
+   * </pre>
    */
-  int getSymbol();
+  long getSymbol();
 }
